+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## Upcoming...
 
 - ... <!-- Add new lines here. Version number will be decided later -->
+- feat: Add a simple `validKennitala()` and `validCreditcard()` checkers
 - feat: `arrayToObject()` now takes "prop" function, emits uniform objects –
   Where for non-object arrays the values of the returned object are now the
   raw array items (Technically a breaking change.)